test_task/requirements.txt
asgiref==3.2.10
Django==3.0.7
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.1.1
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
